+++ ElvUI_JiberishIcons/Init.lua
@@ -21,10 +21,10 @@ JI.Libs = {
 	GUI = _G.LibStub('AceGUI-3.0'),
 }
 
-local GetAddOnMetadata = C_AddOns.GetAddOnMetadata or GetAddOnMetadata
+local GetAddOnMetadata = (C_AddOns and C_AddOns.GetAddOnMetadata) or GetAddOnMetadata
 
 JI.Title = GetAddOnMetadata(AddOnName, 'Title')
-JI.Version = tonumber(GetAddOnMetadata(AddOnName, 'Version'))
+JI.Version = GetAddOnMetadata(AddOnName, 'Version')
 JI.Configs = {}
 JI.myName = UnitName('player')
 JI.myRealm = GetRealmName()
@@ -85,11 +85,12 @@ function JI:MergeStylePacks()
 			end
 		end
 	end
+	if JI.UpdateEllesmereUI then JI:UpdateEllesmereUI() end
 end
 
 local C_AddOns_GetAddOnEnableState = C_AddOns and C_AddOns.GetAddOnEnableState
 local GetAddOnEnableState = GetAddOnEnableState -- eventually this will be on C_AddOns and args swap
-local IsAddOnLoaded = _G.C_AddOns and _G.C_AddOns.IsAddOnLoaded
+local IsAddOnLoaded = (_G.C_AddOns and _G.C_AddOns.IsAddOnLoaded) or _G.IsAddOnLoaded
 local isRetail = WOW_PROJECT_ID == WOW_PROJECT_MAINLINE
 
 function JI:IsAddOnEnabled(addon)